requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-phonenumber-field==7.0.2
django-recaptcha==3.0.0
django-recaptcha3==0.4.0
idna==3.4
phonenumbers==8.13.7
Pillow==9.4.0
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
